Implementing a library manager JSON reader.
The purpose of this thing is to take some LM JSON (which we'll use to
persist the libraries state) and produce the appropriate model
hierarchy, ready for the UI to use.

Had to dump a load of crap in the browser runner page to make the tests
run, this will be dealt with eventually.

There have been some changes to the LM models as well, though these are
probably still quite wrong. More to follow here.