Getting the UI back.
Tee's made various changes since I last looked at the project, and
they've broken the shit out of me. This gets me back on track (ish), and
the new UI rendering.

Oh and setting the production flag to false in the repo version. Ofc
this shouldn't be in here at all, and soon it'll be removed, but for now
yeah.