Implementing coordination between tab groups.
A new 'coordinator' ensures that only one of the registered viewers can
display a tab with a particular id at any one time.

The page load is a mess so it's not all beautiful out of the gate at the
moment, but with a bit of clicking the behaviour can be observed.